moves tests and creates a smoke test suite
